// skinet.client/src/app/api/model/createOrderDTO.ts
/**
 * SKINET.Server
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PaymentSummary } from './paymentSummary';
import { ShippingAddress } from './shippingAddress';


export interface CreateOrderDTO { 
    cartId: string;
    deliveryMethodId: number;
    address: ShippingAddress;
    paymentSummary: PaymentSummary;
}